refactor: CouponForm과 ProductForm 컴포넌트 내부 상태 관리로 개선
